Fixes to gnoll stronghold, adding first draft of shaman calling
While the shaman calling is more or less complete, Lava and I feel it's a bit short so we're adding some additional content before releasing it for testing. However, I didn't want to delay the fixes for the fighter calling in the meantime, hence why it's present but deprecated. I also couldn't exclude it since I reorganized many of the fighter calling files to allow for future expansions.

Gnoll stronghold fixes/changes are better summarized in these two links:
https://www.gibberlings3.net/forums/topic/38385-gnoll-stronghold-v6-review/#comment-339017
https://www.gibberlings3.net/forums/topic/38405-v6-test-playthrough/#comment-339361